backend/requirements.txt
fastapi
uvicorn[standard]
aiosqlite
groq
python-dotenv
pandas
python-multipart
pytest
httpx
